pyproject.toml
[project]
name = "depth-model-survey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "open3d>=0.19.0",
    "pillow>=12.2.0",
    "python-multipart>=0.0.20",
    "torch>=2.11.0",
    "transformers>=5.5.4",
    "uvicorn>=0.37.0",
]
